+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Update minimum ZAP version to 2.10.0.
+- Rename reliability to confidence.
 - standalone/enableDebugLogging.js > use new Log4j 2 APIs.
 - standalone/window_creation_template.js > no longer extend `AbstractFrame`.
 - httpsender/Alert on HTTP Response Code Errors.js and Alert on Unexpected Content Types.js: